requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
redis
celery
django-redis
python-dotenv
Pillow
gunicorn
whitenoise
djangorestframework_simplejwt
drf-yasg
django-filter
stripe